test(http): split http_router_test.exs by concern
Splits the 902-line/39-test mcp_server/test/ptc_runner_mcp/http_router_test.exs
into five concern-scoped files under http/ (auth, host, origin, dispatch,
telemetry), all preserved verbatim. The shared ~20-line setup and conn helpers
move into a new ExUnit.CaseTemplate (PtcRunnerMcp.Http.RouterCase) that each
file uses, so there is no duplicated setup. Every file stays async: false.

Verified: cd mcp_server && mix test runs 39 router tests, 0 failures; mix
precommit green.

Closes #1076
